requirements.txt
selenium==3.141.0
pillow==7.0.0
beautifulsoup4==4.8.2
lxml==4.5.0
requests==2.22.0
Jinja2==2.10.1